requirement.txt
requests
pandas
psycopg2-binary
matplotlib
scikit-learn